Move html_repr implementation into submodule
